pyproject.toml
[tool.poetry]
name = "tachyon"
version = "0.1.0"
description = "Crypto notes with many settings and templates!"
authors = ["Alexander Lavrov <internal@egnod.dev>"]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pydantic = {version = "^1.8.2", extras = ["dotenv"]}
yarl = "^1.6.3"
ujson = "^4.2.0"
tortoise-orm = "^0.17.7"
aerich = "^0.5.8"
asyncpg = "^0.24.0"
aiofiles = "^0.7.0"
httptools = "^0.3.0"
async-exit-stack = "^1.0.1"
async-generator = "^1.10"
fastapi-cache = "^0.1.0"
passlib = "^1.7.4"
argon2-cffi = "^21.1.0"
cryptography = "^35.0.0"
cbor2 = "^5.4.2"
python-rapidjson = "^1.5"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
flake8 = "^3.9.2"
mypy = "^0.910"
isort = "^5.9.3"
yesqa = "^1.2.3"
pre-commit = "^2.11.0"
wemake-python-styleguide = "^0.15.3"
black = "==21.7b0"
autoflake = "^1.4"
pytest-cov = "^2.12.1"
pytest-asyncio = "^0.15.1"
nest-asyncio = "^1.5.1"
pytest-env = "^0.6.2"
requests = "^2.26.0"
asynctest = "^0.13.0"
MonkeyType = "^21.5.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["tachyon",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_return_any = false

[tool.pytest.ini_options]
addopts = "--cov=tachyon"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]
env = [
    "TACHYON_DB_BASE=tachyon_test",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
